📃 chore: AuthenticatedTopPageコンポーネントを削除
不要なコンポーネントだったため。それに伴いAuthenticatedBookItemsコンポーネントも削除。
