+++ src/middleware.ts
@@ -2,6 +2,6 @@ export { auth as middleware } from '@/auth';
 
 export const config = {
   matcher: [
-    '/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$|auth|$|terms|privacy|about|thanks|api/auth/callback/google|users/\\d+).*)',
+    '/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$|auth|$|terms|privacy|about|thanks|api/auth/callback/google|users/+).*)',
   ],
 };
