+++ src/types/user.ts
@@ -1,7 +1,6 @@
 export type UserInfo = {
   handleName: string;
   image: string;
-  id: string;
 };
 
 export type UserParams = {
